ENG-49: Create a new component for selected augment, use augment card to maintain the cards themselves, and make adjustments for the components that use them
