+++ client/app/(protected)/(draft)/(teamBuilder)/players.tsx
@@ -12,11 +12,11 @@ import {
   View,
 } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
-import AugmentCard from "@/components/AugmentCard";
 import FiltersDrawer from "@/components/FiltersDrawer";
 import IconButton from "@/components/IconButton";
 import PlayerData from "@/components/PlayerData";
 import SearchBar from "@/components/SearchBar";
+import SelectedAugment from "@/components/SelectedAugment";
 import Spinner from "@/components/Spinner";
 import Table from "@/components/Table/Table";
 import TeamActionButtons from "@/components/TeamActionButtons";
@@ -206,7 +206,7 @@ const Players = () => {
               />
             </View>
             <View className="gap-4">
-              <AugmentCard />
+              <SelectedAugment />
 
               <TeamBudget />
             </View>
